minor errors
