+++ src/business/UserBusiness.ts
@@ -7,9 +7,6 @@ import { LoginInputDTO } from "../types/LoginInputDTO";
 import { SignupInputDTO } from "../types/SignupInputDTO"
 import { InvalidInputError } from "./errors/InvalidInputError";
 import { NotFoundError } from "./errors/NotFoundError";
-import { DAY, MarcarShow } from "../types/MarcarShow";
-import { CustomError } from "./errors/CustomError";
-import { SetShow } from "../model/SetShow";
 
 export class UserBusiness {
     constructor(
@@ -75,48 +72,5 @@ export class UserBusiness {
         })
 
         return token
-    }
-
-    setShow = async (dia: MarcarShow) => {
-        const { day, startingTime, endingTime, id } = dia
-
-        if(!day || !startingTime || !endingTime || !id) {
-            throw new InvalidInputError("Invalid input. day and time are required")
-        }
-
-        if(day !== DAY.FRIDAY && day !== DAY.SATURDAY && day !== DAY.SUNDAY){
-            throw new CustomError(400, "day must be FRIDAY, SUNDAY OR SATURDAY")
-        }
-
-        if (startingTime < 8 || startingTime > 22) {
-            throw new CustomError(400, "The starting time must be between 8h and 22h")
-        }
-
-        if (endingTime < 9 || endingTime > 23) {
-            throw new CustomError(400, "The ending time must be between 9 and 23")
-        }
-
-        const showFromDB = await this.userDatabase.getShowByDayAndTime(day)
-
-        showFromDB.forEach((show)=>{
-            if(show.start_time<startingTime && show.end_time>startingTime){
-                throw new CustomError(409, `Starting time not available, ${show.start_time}:00 - ${show.end_time}:00`)
-            }
-            if(show.start_time<endingTime && show.end_time>endingTime){
-                throw new CustomError(409, `Ending time not available, ${show.start_time}:00 - ${show.end_time}:00`)
-            }
-            if(show.start_time>startingTime&&show.start_time<endingTime){
-                throw new CustomError(409, `Ending time not available, another show starts before the wanted ending time, ${show.start_time}:00 - ${show.end_time}:00`)
-            }
-            if(show.end_time>startingTime&&show.end_time<endingTime){
-                throw new CustomError(409, `Time not available, ${show.start_time}:00 - ${show.end_time}:00`)
-            }
-        })
-
-        const showId = IdGenerator.idGenerator()
-
-        const show = new SetShow(showId, day, startingTime, endingTime, id)
-
-        await this.userDatabase.insertShow(show)
     }
 }
