+++ src/controller/UserController.ts
@@ -41,16 +41,4 @@ export class UserController {
             res.status(error.statusCode || 500).send({ message: error.message })
         }
     }
-
-    marcarShow = async (req: Request, res: Response) => {
-        try {
-            const { day, startingTime, endingTime, id } = req.body
-
-            await this.userBusiness.setShow({day, startingTime, endingTime, id})
-
-            res.send(`Show registered: ${day}, from ${startingTime}:00 to ${endingTime}:00`)
-        } catch (error: any) {
-            res.send(error.sqlMessage || error.message)
-        }
-    }
 }
