+++ src/routes/userRouter.ts
@@ -16,4 +16,3 @@ export const userRouter = express.Router()
 
 userRouter.post('/signup', userController.signUp)
 userRouter.post('/login', userController.login) 
-userRouter.post('/show/register', userController.marcarShow)
